+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## [1.0.0] 
+## [1.0.0] - [2018-07-01]
 ### Added 
 - Created and finalized the static website and the blog. Next is to setup different environment.
 
 ### Changed
 - Removed TD mention from our front page because we haven't heard back from TD.
+- Refactored Press page components into /components/Press
+- Updated styles on index and press pages.